requirements.txt
fastapi
uvicorn
cryptography
pillow
numpy
python-multipart

.venv\Scripts\activate

uvicorn main:app --reload
# This command activates the virtual environment and starts the FastAPI application with hot-reloading enabled. Make sure to run this command in the terminal while being in the project directory where the `main.py` file is located.

cd frontend
python -m http.server 5500
# This command changes the current directory to the `frontend` folder and starts a simple HTTP server on port 5500. You can access the frontend by navigating to `http://localhost:5500` in your web browser.